+++ frontend/src/pages/MyPage.style.js
@@ -1,4 +1,3 @@
-// MyPage.styles.js
 import styled from "styled-components";
 
 export const Container = styled.div`
